[NPCData] Improve dialogue pool pull logic
