--- Assets/NPCDataSystem-Josie/NPCData.asset
@@ -1,16 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: ed558470323453347a9e3ab4d78cfbf7, type: 3}
-  m_Name: NPCData
-  m_EditorClassIdentifier: Assembly-CSharp::NPCData
-  npcName: Josephine
-  npcPotraitID: {fileID: 2800000, guid: 727a75301c3d24613a3ebcec4a24c2c8, type: 3}
