--- Assets/NPCDataSystem-Josie/NPCData.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c909476e8b6f30441afc1af2a3781041
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
