+++ Assets/ScriptableObjects/Dialogue/_NpcDisplayInfo/Sam.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0e39476b7d58709449cea31648de0af6
+guid: 2b9034908e65964479bb08d9fd99effb
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
