+++ Assets/Scripts/Systems/NPCDataSystem-Josie/DialogueTrigger.cs
