+++ Assets/Scripts/Systems/NPCDataSystem-Josie/NPCData.cs.meta
