+++ Assets/Scripts/Systems/NPCDataSystem-Josie/TownUI.cs
@@ -1,0 +1,16 @@
+using UnityEngine;
+
+public class TownUI : MonoBehaviour
+{
+    // Start is called once before the first execution of Update after the MonoBehaviour is created
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+}
