+++ Assets/Scripts/Systems/NPCDataSystem-Josie/TownUI.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 341cda04b7f0040459fc9924615015e5
