+++ Assets/Scripts/Systems/NPCDataSystem-Josie/empty.prefab
